feat: add ServUO assistant
